Increased max vertex count

+++ src/render/vertex.rs
@@ -7,5 +7,5 @@ pub struct Vertex {
 
 pub struct Mesh {
     pub vertices: Vec<Vertex>,
-    pub indices: Vec<u16>,
+    pub indices: Vec<u32>,
 }
